requirements.txt
numpy
opencv-python
scikit-learn
scikit-image
streamlit
Pillow

